requirements.txt
pandas>=2.1
xlrd>=2.0.1
python-dateutil>=2.8
SQLAlchemy>=2.0
psycopg2-binary>=2.9
python-dotenv>=1.0
streamlit>=1.37
plotly>=5.22
pytest>=8.0
requests>=2.32
streamlit==1.39.0
plotly==5.24.1
sqlalchemy==2.0.35
pg8000==1.31.2
python-dotenv==1.0.1